backend/backend_api/api/requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
jedi==0.19.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mongoengine==0.28.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pudb==2024.1.2
pycairo==1.16.2
pycups==1.9.73
pygments==2.18.0
PyGObject==3.36.0
PyJWT==2.1.0
pymacaroons==0.13.0
pymongo==3.11.4
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.19.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
systemd-python==234
ttystatus==0.38
typing-extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
urwid==2.6.15
urwid-readline==0.14
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.13
xkit==0.0.0
zipp==1.0.0
